backend/requirements.txt
aerich==0.5.5
asyncpg==0.23.0
fastapi==0.68.0
tortoise-orm==0.17.6
uvicorn==0.14.0
bcrypt==3.2.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5